add preprocess cuda kernel
