+++ onnxruntime/contrib_ops/cuda/quantization/matmul_nbits.cc
@@ -126,18 +126,19 @@ Status MatMulNBits<T>::PrePack_B([[maybe_unused]] const Tensor& tensor,
           stream, packed_transposed_weight, blob_data, n, k);
     }
 
+    using onnxruntime::llm::kernels::cutlass_kernels::QuantType;
+    QuantType quant_type = nbits_ == 4 ? QuantType::W4_A16 : QuantType::W8_A16;
+
+#ifdef FPA_INTB_PREPROCESS_IN_CPU
     CUDA_RETURN_IF_ERROR(cudaStreamSynchronize(stream));
 
     auto tranpose_weight_buffer = this->AllocateBufferOnCPUPinned<int8_t>(packed_weight_bytes);
     CUDA_RETURN_IF_ERROR(cudaMemcpy(tranpose_weight_buffer.get(), packed_transposed_weight, packed_weight_bytes, cudaMemcpyDeviceToHost));
 
     auto processed_weight_buffer = this->AllocateBufferOnCPUPinned<uint8_t>(n * k / (8 / nbits_));
-    bool force_interleave = false;
-
-    using onnxruntime::llm::kernels::cutlass_kernels::QuantType;
-    QuantType quant_type = nbits_ == 4 ? QuantType::W4_A16 : QuantType::W8_A16;
 
     // TODO: Add a cuda kernle for preprocessing so that we can avoid copying the data back to CPU.
+    constexpr bool force_interleave = false;
     onnxruntime::llm::kernels::cutlass_kernels::preprocess_weights_for_mixed_gemm(
         reinterpret_cast<int8_t*>(processed_weight_buffer.get()),
         reinterpret_cast<const int8_t*>(tranpose_weight_buffer.get()),
@@ -146,8 +147,19 @@ Status MatMulNBits<T>::PrePack_B([[maybe_unused]] const Tensor& tensor,
         force_interleave);
 
     CUDA_RETURN_IF_ERROR(cudaMemcpy(preprocessed_weight, processed_weight_buffer.get(), n * k / (8 / nbits_), cudaMemcpyHostToDevice));
-    CUDA_RETURN_IF_ERROR(cudaDeviceSynchronize());
+#else
+    auto permutation_map_buffer = this->GetTransientScratchBuffer<int32_t>(32);
+    onnxruntime::llm::kernels::cutlass_kernels::preprocess_weights_for_mixed_gemm_cuda(
+        stream,
+        sm_,
+        preprocessed_weight,
+        packed_transposed_weight,
+        permutation_map_buffer.get(),
+        {static_cast<size_t>(k), static_cast<size_t>(n)},
+        quant_type);
+#endif
 
+    CUDA_RETURN_IF_ERROR(cudaDeviceSynchronize());
     DUMP_TENSOR_INIT();
     DUMP_TENSOR_D("packed transposed_weight in GPU", packed_transposed_weight, k, n * nbits_ / 8);
     DUMP_TENSOR_D("preprocessed_weight", reinterpret_cast<uint8_t*>(preprocessed_weight), k, n * nbits_ / 8);
